requirements.txt
# ── Core ──────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.1

# ── Auth ──────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.1

# ── Redis / Celery ────────────────────────────
celery>=5.4.0
redis>=5.0.4
flower>=2.0.1

# ── HTTP / API ────────────────────────────────
httpx>=0.27.0

# ── GitHub ────────────────────────────────────
PyGithub>=2.3.0
GitPython>=3.1.43

# ── Analyzers ─────────────────────────────────
flake8>=7.0.0
pylint>=3.1.0
bandit>=1.7.8
pip-audit>=2.7.3

# ── Utils ─────────────────────────────────────
python-dotenv>=1.0.1
pydantic[email]>=2.7.1
requests>=2.32.2
